\name{dt}
\docType{data}
\alias{dt}
\title{Ocean Bottom Seismometer Sample Rate}
\description{
This is the sample rate for the instrument that recorded \code{\link{sig}}.
}
\usage{data(port_foster_event)}
\format{
A float describing the sample rate (1/second).
}
\source{Ocean bottom seismometer records from the 2005 TOMODEC active source tomography experiment, Deception Island, Antarctica.}
\keyword{datasets}
