\name{evolutive_fft}
\alias{evolutive_fft}
\title{Generate Fourier spectrogram}
\description{
This function generates a spectrogram using the evolutive FFT method.
It is a modified copy of the \code{evolfft} function in the RSEIS package, and still depends on this package to work.
}
\usage{
evolutive_fft(sig, dt, ft, freq_span, taper = 0.05) 
}
\arguments{
\item{sig}{The signal to process}
\item{dt}{sample rate}
\item{ft}{Fourier spectrogram options}
\itemize{
    \item{\code{ft$nfft} is the fft length}
    \item{\code{ft$ns} is the number of samples in a window}
    \item{\code{ft$nov} is the number of samples to overlap}}
\item{freq_span}{Frequency span to render spectrogram over.  \code{c(0, -1)} plots everything up to the Nyquist frequency.}
\item{taper}{Taper value to use for spectrogram (default is 0.05)}
}

\details{
This function is a simple Fourier spectrogram plotter.
It's useful to compare this image with images generated by \code{\link{hhspec_image}} to see how the Fourier and Hilbert spectrograms differ.
}

\references{Jonathan M. Lees (2012). RSEIS: Seismic Time Series Analysis Tools. R
  package version 3.0-6. http://CRAN.R-project.org/package=RSEIS}

\author{Daniel Bowman \email{daniel.bowman@unc.edu}, Jonathan M. Lees}

\seealso{
\code{\link{ftspec_image}}
}
\examples{
data(port_foster_event)

ft=list()
ft$nfft=4096
ft$ns=30
ft$nov=29

freq_span=c(0, 25)
ev = evolutive_fft(sig, dt, ft, freq_span)

#Plot raw spectrogram
image_z=t(ev$DSPEC[1:(ev$numfreqs/2),])
f_ind=(ev$freqs>=freq_span[1] & ev$freqs <= freq_span[2])
image_z=t(ev$DSPEC[1:(ev$numfreqs/2),])
image_z=image_z[,f_ind]
image(image_z)
}

\keyword{nonparametric}
