\name{hh_spectrum}
\alias{hh_spectrum}
\title{Generate Hilbert spectrum}
\description{
Generates a Hilbert periodogram from the results of \code{\link{sig2imf}} and \code{\link{eemd}}.
}
\usage{
hh_spectrum(hres, dfreq, freq_span = NULL, time_span = NULL, 
    scaling = "none", verbose = TRUE)
}
\arguments{
\item{hres}{This is the output generated by \code{\link{eemd_compile}} or \code{\link{eemd_resift}}}
\item{dfreq}{Frequency resolution of spectrum}
\item{time_span}{Time span to render spectrum over; \code{NULL} means over the whole time series}
\item{freq_span}{Frequency span to include in spectrum; \code{NULL} means render all the frequencies in the time series}
\item{scaling}{If \code{"log"}, render a log10 frequency spectrum.  Defaults to \code{"none"} (linear).}
\item{verbose}{If \code{TRUE}, print progress messages}
}
\details{
\code{hh_spectrum} sums Hilbert spectral data over the time domain to produce the equivalent of a periodogram.  
The result can be plotted using \code{\link{hhspec_plot}}.
}
\value{ 
\item{hspec}{A data structure containing the spectrum of each IMF.}
}

\author{Daniel Bowman \email{daniel.bowman@unc.edu}}

\seealso{
\code{\link{hh_render}}, \code{\link{hhspec_plot}}
}
\examples{
data(port_foster_event)

emd_result <- sig2imf(sig, tt)

dfreq <- 0.1
hspec <- hh_spectrum(emd_result, dfreq)
hhspec_plot(hspec, show_fourier = TRUE, scale_fourier = TRUE)

}
\keyword{ts}
