\name{sig}
\docType{data}
\alias{sig}
\title{Transitory Seismic Event at Deception Island Volcano}
\description{
This is 20 seconds of data from the 2005 TOMODEC ocean bottom seismometer network at Deception Island, South Shetland Islands, Antarctica with sample rate \code{\link{tt}}.
It shows one of several thousand transitory seismic events occurring in Port Foster (the flooded caldera of the volcano).
}
\usage{data(PortFosterEvent)}
\format{
A 2500 element vector containing the seismic record.  Units are meters per second.
}
\source{Ocean bottom seismometer records from the 2005 TOMODEC active source tomography experiment, Deception Island, Antarctica.}
\keyword{datasets}
