\name{tt}
\docType{data}
\alias{tt}
\title{Ocean Bottom Seismometer Sample Rate}
\description{
This is the sample times for the instrument that recorded \code{\link{sig}}.
}
\usage{data(PortFosterEvent)}
\format{
A vector describing the sample times. The sample rate was constant at 125 samples per second.
}
\source{Ocean bottom seismometer records from the 2005 TOMODEC active source tomography experiment, Deception Island, Antarctica.}
\keyword{datasets}
