\name{ftspec_image}
\alias{ftspec_image}
\title{Display Fourier spectrogram}
\description{
This function displays a Fourier spectrogram using the same plot structure and options as \code{\link{hhgram_image}}.
It uses the function \code{\link{evolutive_fft}} to generate a spectrogram, then wraps it in the same plotting format as \code{\link{hhgram_image}}.
}
\usage{
ftspec_image(sig, dt, ft, time_span = NULL, freq_span = NULL, amp_span = NULL,
     taper=0.05, scaling = "none", grid=TRUE, colorbar=TRUE, 
     backcol=c(0, 0, 0), colormap=NULL, pretty=FALSE, ...)
}
\arguments{
\item{sig}{The signal to process}
\item{dt}{sample rate}
\item{ft}{Fourier spectrogram options}
\itemize{
    \item{\code{ft$nfft} is the fft length}
    \item{\code{ft$ns} is the number of samples in a window}
    \item{\code{ft$nov} is the number of samples to overlap}}
\item{time_span}{Time span to render spectrogram over.  \code{NULL} will draw the spectrogram over the entire signal.}
\item{freq_span}{Frequency span to render spectrogram over.  \code{NULL} plots everything up to the Nyquist frequency.}
\item{amp_span}{Amplitude range to plot.  \code{NULL} plots everything.}
\item{taper}{Taper value to use for spectrogram (default is 0.05), see \code{\link{cosine_taper}}.}
\item{scaling}{determines whether to apply logarithmic (log), or square root (sqrt) scaling to the amplitude data}
\item{grid}{Boolean - whether to display grid lines or not}
\item{colorbar}{Boolean - whether to display amplitude colorbar or not}
\item{backcol}{What background color to use behind the spectrogram, in a 3 element vector: \code{c(red, green, blue)}}
\item{colormap}{What palette object to use for the spectrogram, defaults to \code{rainbow(colorbins,start=0,end=5/6)}}
\item{pretty}{Boolean - to choose nice axes values or to use exactly the ranges given}
\item{This function supports some optional parameters as well:}
\itemize{
    \item{trace_format}{the format of the trace minima and maxima in sprintf format}
    \item{img_x_format}{the format of the X axis labels of the image in sprintf format}
    \item{img_y_format}{the format of the Y axis labels of the image in sprintf format}
    \item{colorbar_format}{the format of the colorbar labels in sprintf format}
    \item{cex.lab}{is the font size of the image axis labels}
    \item{cex.colorbar}{the font size of the colorbar}
    \item{cex.trace}{the font size of the trace axis labels}
    \item{img_x_lab}{the X - axis label of the image, it defaults to "time"}
    \item{img_y_lab}{the Y - axis label of the image, it defaults to "frequency"}
    \item{main}{Adds a title to the figure}}
}

\details{
This function is a simple Fourier spectrogram plotter.
It's useful to compare this image with images generated by \code{\link{hhgram_image}} to see how the Fourier and Hilbert spectrograms differ.
}
\value{
\item{img}{The spectrogram image, suitable for plotting with the generic \code{image} function}
}
\references{Jonathan M. Lees (2012). RSEIS: Seismic Time Series Analysis Tools. R
  package version 3.0-6. http://CRAN.R-project.org/package=RSEIS}

\author{Daniel Bowman \email{daniel.bowman@unc.edu}}

\seealso{
\code{\link{hhgram_image}}, \code{\link{evolutive_fft}}
}
\examples{
data(port_foster_event)

dt = mean(diff(tt))

ft=list()
ft$nfft=4096
ft$ns=30
ft$nov=29

time_span=c(5, 10)
freq_span=c(0, 25)
amp_span=c(1e-5, 0.0003)
ftspec_image(sig, dt, ft, 
    time_span = time_span, freq_span = freq_span, amp_span = amp_span, 
    pretty = TRUE, img_x_format = "%.1f", img_y_format = "%.0f",
    main = "Port Foster Event - Fourier Spectrogram")
}

\keyword{nonparametric}
